refactor into reward pallet
