--- frame/reward-curve/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "pallet-reward-curve"
-version = "2.3.0"
-authors = ["Shawn Tabrizi <shawntabrizi@gmail.com>"]
-license = "GPL-3.0-or-later"
-edition = "2018"
-description = "Control block rewards via a reward curve."
-
-[dependencies]
-serde = { version = "1.0.101", optional = true }
-codec = { package = "parity-scale-codec", version = "1.3.0", default-features = false, features = ["derive"] }
-sp-std = { path = "../../vendor/substrate/primitives/std", default-features = false }
-sp-runtime = { path = "../../vendor/substrate/primitives/runtime", default-features = false }
-frame-system = { path = "../../vendor/substrate/frame/system", default-features = false }
-frame-support = { path = "../../vendor/substrate/frame/support", default-features = false }
-pallet-rewards = { path = "../rewards", default-features = false }
-
-[dev-dependencies]
-sp-core = { path = "../../vendor/substrate/primitives/core", default-features = false }
-sp-io = { path = "../../vendor/substrate/primitives/io", default-features = false }
-pallet-balances = { path = "../../vendor/substrate/frame/balances", default-features = false }
-sp-consensus-pow = { path = "../../vendor/substrate/primitives/consensus/pow", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-	"serde",
-	"codec/std",
-	"sp-std/std",
-	"sp-runtime/std",
-	"frame-system/std",
-	"frame-support/std",
-	"pallet-rewards/std",
-]
